switch tile x and y to i32

+++ migrations/2024-11-01-080726_tile_cache/up.sql
@@ -2,8 +2,8 @@
 CREATE TABLE gtfs.tile_storage (
     category smallint not null,
     z smallint not null,
-    x smallint not null,
-    y smallint not null,
+    x int not null,
+    y int not null,
     mvt_data bytea not null,
     added_time Timestamptz not null,
